feat: extract motion data when click on build button, fix clippy
